requirements.txt
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-oauth2-authcodeflow==0.8.1
django-rest-authemail==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.4
GitLab.py==0.112
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
gunicorn==20.1.0
py3dns==3.2.1
pymongo==3.12.3
python-gitlab==3.12.0
validate-email==1.3
whitenoise==6.3.0
python-decouple==3.7
django-postgres-copy==2.3.1
psycopg2==2.8.6
python-dateutil==2.7.0
django_heroku==0.3.1
dj-database-url==1.2.0
psycopg2-binary==2.9.5
google-api-python-client==1.31.5
google-auth-oauthlib==1.0.0
redis==4.5.1
spotipy==2.22.1